Fixed issue with not being able to enter decimal numbers in number fields.
git-svn-id: http://archfirst.googlecode.com/svn/trunk@188 5f1d654b-2d44-f8f1-813d-ae2f855fe689

+++ dotnet/Projects/Bullsfirst/Silverlight/Bullsfirst.Module.Transfer/ViewModels/TransferViewModel.cs
@@ -376,8 +376,8 @@ namespace Bullsfirst.Module.Transfer.ViewModels
             }
         }
 
-        private decimal _quantity;
-        public decimal Quantity
+        private int _quantity;
+        public int Quantity
         {
             get { return _quantity; }
             set
